+++ mockwebserver/src/main/java/okhttp3/mockwebserver/MockWebServer.java
@@ -57,7 +57,7 @@ import okhttp3.HttpUrl;
 import okhttp3.Protocol;
 import okhttp3.Request;
 import okhttp3.Response;
-import okhttp3.Streams;
+import okhttp3.internal.ws.Streams;
 import okhttp3.internal.Internal;
 import okhttp3.internal.NamedRunnable;
 import okhttp3.internal.Util;
