+++ okhttp-tests/src/test/java/okhttp3/internal/ws/RealWebSocketTest.java
@@ -24,7 +24,6 @@ import java.util.concurrent.TimeUnit;
 import okhttp3.Protocol;
 import okhttp3.Request;
 import okhttp3.Response;
-import okhttp3.Streams;
 import okio.ByteString;
 import okio.Okio;
 import okio.Pipe;
