+++ okhttp/src/main/java/okhttp3/internal/ws/RealWebSocket.java
@@ -16,11 +16,8 @@
 package okhttp3.internal.ws;
 
 import java.io.IOException;
-import java.net.ProtocolException;
 import java.net.SocketTimeoutException;
 import java.util.ArrayDeque;
-import java.util.Collections;
-import java.util.List;
 import java.util.Random;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
@@ -28,18 +25,11 @@ import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.Nullable;
 import okhttp3.Call;
-import okhttp3.Callback;
-import okhttp3.EventListener;
-import okhttp3.OkHttpClient;
-import okhttp3.Protocol;
 import okhttp3.Request;
 import okhttp3.Response;
-import okhttp3.Streams;
 import okhttp3.WebSocket;
 import okhttp3.WebSocketListener;
-import okhttp3.internal.Internal;
 import okhttp3.internal.Util;
-import okhttp3.internal.connection.StreamAllocation;
 import okio.BufferedSink;
 import okio.ByteString;
 import okio.Okio;
